test: Move to XCTest from dependency based
