+++ .gitignore
@@ -6,6 +6,7 @@
 ## Build generated
 build/
 DerivedData
+.build/
 
 ## Various settings
 *.pbxuser
