--- Cartfile.private
@@ -1,2 +1,0 @@
-github "Quick/Quick" ~> 2.0.0
-github "Quick/Nimble" ~> 8.0.0
