+++ Tests/MessageKitTests/Mocks/MockMessage.swift
@@ -1,7 +1,7 @@
 /*
  MIT License
 
- Copyright (c) 2017-2019 MessageKit
+ Copyright (c) 2017-2020 MessageKit
 
  Permission is hereby granted, free of charge, to any person obtaining a copy
  of this software and associated documentation files (the "Software"), to deal
@@ -23,6 +23,7 @@
  */
 
 import Foundation
+import UIKit
 import CoreLocation
 import AVFoundation
 @testable import MessageKit
