+++ Tests/MessageKitTests/Mocks/MockMessagesDataSource.swift
@@ -1,18 +1,18 @@
 /*
  MIT License
- 
- Copyright (c) 2017-2019 MessageKit
- 
+
+ Copyright (c) 2017-2020 MessageKit
+
  Permission is hereby granted, free of charge, to any person obtaining a copy
  of this software and associated documentation files (the "Software"), to deal
  in the Software without restriction, including without limitation the rights
  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  copies of the Software, and to permit persons to whom the Software is
  furnished to do so, subject to the following conditions:
- 
+
  The above copyright notice and this permission notice shall be included in all
  copies or substantial portions of the Software.
- 
+
  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -22,29 +22,31 @@
  SOFTWARE.
  */
 
-import XCTest
+import Foundation
 @testable import MessageKit
 
-class MessagesCollectionViewTests: XCTestCase {
+class MockMessagesDataSource: MessagesDataSource {
 
-    var messagesCollectionView: MessagesCollectionView!
-    let rect = CGRect(x: 0, y: 0, width: 100, height: 100)
-    let layout = MessagesCollectionViewFlowLayout()
+    var messages: [MessageType] = []
+    let senders: [MockUser] = [
+        MockUser(senderId: "sender_1", displayName: "Sender 1"),
+        MockUser(senderId: "sender_2", displayName: "Sender 2")
+    ]
 
-    override func setUp() {
-        super.setUp()
-        messagesCollectionView = MessagesCollectionView(frame: rect, collectionViewLayout: layout)
+    var currentUser: MockUser {
+        return senders[0]
     }
 
-    override func tearDown() {
-        messagesCollectionView = nil
-        super.tearDown()
+    func currentSender() -> SenderType {
+        return currentUser
     }
 
-    func testInit() {
-        XCTAssertEqual(messagesCollectionView.frame, rect)
-        XCTAssertEqual(messagesCollectionView.collectionViewLayout, layout)
-        XCTAssertEqual(messagesCollectionView.backgroundColor, UIColor.collectionViewBackground)
+    func numberOfSections(in messagesCollectionView: MessagesCollectionView) -> Int {
+        return messages.count
+    }
+
+    func messageForItem(at indexPath: IndexPath, in messagesCollectionView: MessagesCollectionView) -> MessageType {
+        return messages[indexPath.section]
     }
 
 }
