+++ Tests/MessageKitTests/Model Tests/AvatarTests.swift
@@ -1,6 +1,3 @@
-// swift-tools-version:5.3
-// The swift-tools-version declares the minimum version of Swift required to build this package.
-
 /*
  MIT License
 
@@ -12,8 +9,10 @@
  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  copies of the Software, and to permit persons to whom the Software is
  furnished to do so, subject to the following conditions:
+
  The above copyright notice and this permission notice shall be included in all
  copies or substantial portions of the Software.
+
  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -23,30 +22,14 @@
  SOFTWARE.
  */
 
-import PackageDescription
+import Foundation
+import XCTest
+@testable import MessageKit
 
-let package = Package(
-    name: "MessageKit",
-    platforms: [.iOS(.v12)],
-    products: [
-        .library(name: "MessageKit", targets: ["MessageKit"]),
-    ],
-    dependencies: [
-        .package(url: "https://github.com/nathantannar4/InputBarAccessoryView", .upToNextMajor(from: "5.2.1"))
-    ],
-    targets: [
-        .target(
-            name: "MessageKit",
-            dependencies: ["InputBarAccessoryView"],
-            path: "Sources",
-            exclude: ["Supporting/Info.plist", "Supporting/MessageKit.h"],
-            swiftSettings: [SwiftSetting.define("IS_SPM")]
-        ),
-        .testTarget(
-            name: "MessageKitTests",
-            dependencies: ["MessageKit"],
-            exclude: ["Resources/sound1.m4a", "Supporting Files/Info.plist"]
-        )
-    ],
-    swiftLanguageVersions: [.v5]
-)
+final class AvatarTests: XCTestCase {
+    func testAvatarDefaultPropertyValues() {
+        let avatar = Avatar()
+        XCTAssertNil(avatar.image)
+        XCTAssertEqual(avatar.initials, "?")
+    }
+}
