+++ Tests/MessageKitTests/Protocols Tests/MessagesDisplayDelegateTests.swift
@@ -1,7 +1,7 @@
 /*
  MIT License
 
- Copyright (c) 2017-2019 MessageKit
+ Copyright (c) 2017-2020 MessageKit
 
  Permission is hereby granted, free of charge, to any person obtaining a copy
  of this software and associated documentation files (the "Software"), to deal
@@ -22,11 +22,11 @@
  SOFTWARE.
  */
 
+import Foundation
 import XCTest
-
 @testable import MessageKit
 
-class MessagesDisplayDelegateTests: XCTestCase {
+final class MessagesDisplayDelegateTests: XCTestCase {
 
     private var sut: MockMessagesViewController!
