+++ Tests/MessageKitTests/Views Tests/MessagesCollectionViewTests.swift
@@ -1,7 +1,7 @@
 /*
  MIT License
 
- Copyright (c) 2017-2019 MessageKit
+ Copyright (c) 2017-2020 MessageKit
 
  Permission is hereby granted, free of charge, to any person obtaining a copy
  of this software and associated documentation files (the "Software"), to deal
@@ -22,28 +22,29 @@
  SOFTWARE.
  */
 
-import Quick
-import Nimble
+import XCTest
 @testable import MessageKit
 
-final class SenderSpec: QuickSpec {
+class MessagesCollectionViewTests: XCTestCase {
 
-    override func spec() {
-        describe("equality between two Senders") {
-            context("they have the same id ") {
-                it("should be equal") {
-                    let sender1 = MockUser(senderId: "1", displayName: "Steven")
-                    let sender2 = MockUser(senderId: "1", displayName: "Nathan")
-                    expect(sender1.senderId == sender2.senderId).to(equal(true))
-                }
-            }
-            context("they have a different id") {
-                it("should not be equal") {
-                    let sender1 = MockUser(senderId: "1", displayName: "Steven")
-                    let sender2 = MockUser(senderId: "2", displayName: "Nathan")
-                    expect(sender1.senderId == sender2.senderId).to(equal(false))
-                }
-            }
-        }
+    var messagesCollectionView: MessagesCollectionView!
+    let rect = CGRect(x: 0, y: 0, width: 100, height: 100)
+    let layout = MessagesCollectionViewFlowLayout()
+
+    override func setUp() {
+        super.setUp()
+        messagesCollectionView = MessagesCollectionView(frame: rect, collectionViewLayout: layout)
     }
+
+    override func tearDown() {
+        messagesCollectionView = nil
+        super.tearDown()
+    }
+
+    func testInit() {
+        XCTAssertEqual(messagesCollectionView.frame, rect)
+        XCTAssertEqual(messagesCollectionView.collectionViewLayout, layout)
+        XCTAssertEqual(messagesCollectionView.backgroundColor, UIColor.collectionViewBackground)
+    }
+
 }
